tests: organize behave steps
